Book_Store_BE/staff-service/requirements.txt
django
djangorestframework
requests
